Move the ahp page under admin
